feat(http): Add Web Standard HTTP handler with Node.js adapter
- Add createHandler for Web Standard (Request) => Response pattern
- Add toNodeHandler adapter for Node.js http.createServer
- Works natively in Deno, Bun, and Cloudflare Workers
- Node.js requires the adapter due to lack of native Web Standard server
- Add feathers/http/node export for Node.js adapter
- Add documentation for all runtimes
